Day 10 part 2.
Finally!

1. Approach 1:

A top-down recursive approach where I determine, for each entry in the
input data, whether its removal yields a valid adaptor chain, and if so,
I construct a vector where this element is missing and apply the
counting to the resulting vector. This has a lot of allocations in
consequence which is the first inefficiency.

Another error I did at first was to not consider the charging outlet
with index 0. Therefore the input 1,2,3,4 gave only chains starting
with 1, e.g. 1,3,4 or 1,2,4 while 0,2,3,4 and 0,3,4 are also valid.

2. Approach 2:
To get rid of allocating new vectors, I changed to a bottom-up recursive
approach, where instead of eleminating adaptors from the chain, I count
the possibilities there are for a adding adaptors starting from a given
adaptor. This was still way too slow so I had to add memoization. I
first did the exercise in Python because I suspected that introducing a
mutable reference to a HashMap in a recursive function might lead to
problems with the borrow checker. When the Python version turned out to
be very fast, I also knew that it was the right solution. Implementing
it in Rust turned out to be quite easy, although if it very necessary
to parallelize this with rayon some more work would be required.