Day 23, Task 2.
This took me a very long time to solve. All my attempts were very
slow and took forever. Unfortunately, I had to look for hints on
Reddit, where I learned that the solution can be coded up a lot more
"locally" without having to allocate a lot of memory.

With this approach, the solution is fast.